Towards #2107 

Major changes:

- use native readable streams optimization. `Deno.File#readable` instead of a wrapper stream. See https://github.com/denoland/deno/pull/14284
- use fnv-1a default etag hash function. Adds an option to allow using others.
- ~~faster hex function for out of other algorithms (like `sha-1`), using hex lookup table.~~
- disables logging by default when invoked independently. Replaces the `--quiet` flag with `--verbose`. Public API remains unchanged.
- ~~use simpler extname extraction to avoid scanning manually. (downsides?)~~

Benchmarks:

`main`:
```
Running 10s test @ http://localhost:4507/README.md
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.44ms  522.77us  26.73ms   97.55%
    Req/Sec     1.40k   119.07     2.63k    97.39%
  112288 requests in 10.10s, 432.42MB read
Requests/sec:  11114.28
Transfer/sec:     42.80MB
```

This patch:
```
wrk -t8 -c20 -d10s http://localhost:4507/README.md
Running 10s test @ http://localhost:4507/README.md
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.01ms  163.42us   6.32ms   85.73%
    Req/Sec     1.99k    67.51     2.09k    86.51%
  160068 requests in 10.10s, 611.53MB read
Requests/sec:  15848.72
Transfer/sec:     60.55MB
```